collision with walls

+++ cmp/PhysicsComponent.hpp
@@ -2,7 +2,7 @@
 #include <SFML/System/Vector2.hpp>
 struct PhysicsComponent
 {
-	sf::Vector2f vel{};
+	sf::Vector2i vel{};
 	sf::Vector2f pos{};
 	PhysicsComponent(){}
 	static const int mask = 0b10;
